pyproject.toml
[project]
name = "layout-prompter"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Shunsuke KITADA", email = "shunsuke.kitada.0831@gmail.com" },
]
requires-python = ">=3.11"
dependencies = [
    "datasets[vision]>=2.0.0",
    "langchain>=0.3.0",
    "loguru>=0.7.0",
    "opencv-python>=4.11.0.86",
    "pydantic-numpy>=5.0.2",
    "pydantic-settings>=2.8.1",
]

[project.optional-dependencies]
openai = ["langchain-openai>=0.3.11"]
aws = ["langchain-aws>=0.2.18"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "ipywidgets>=8.1.6",
    "jupyterlab>=4.4.0",
    "mypy>=1.0.0",
    "pytest>=6.0.0",
    "pytest-cov>=6.0.0",
    "ruff>=0.1.5",
]

[tool.mypy]
python_version = "3.11"
ignore_missing_imports = true
exclude = "tests"
plugins = ['pydantic.mypy']
